Move `detail_response` function
